improve structure & url list logic
