+++ Stanislav_Dzisiak/2/Gemfile
@@ -4,6 +4,7 @@ source 'https://rubygems.org'
 
 gemspec
 
+gem 'celluloid', '~> 0.18.0'
 gem 'faraday', '~> 1.4.0'
 gem 'rake', '~> 13.0'
 gem 'slop', '~> 4.8.2'
