+++ Stanislav_Dzisiak/2/url_analyzer.gemspec
@@ -25,6 +25,7 @@ Gem::Specification.new do |spec|
   spec.executables   = spec.files.grep(%r{\Aexe/}) { |f| File.basename(f) }
   spec.require_paths = ['lib']
 
+  spec.add_dependency 'celluloid', '~> 0.18.0'
   spec.add_dependency 'faraday', '~> 1.4.0'
   spec.add_dependency 'slop', '~> 4.8.2'
 end
